implemented improvements and fixed issues highlighted in the review
