+++ server/@types/index.d.ts
@@ -50,3 +50,38 @@ export type RdsEntry = {
   is_production?: string
   environment_name?: string
 }
+
+interface AlertListResponseDataItem {
+  annotations?: {
+    dashboard_url?: string
+    message?: string
+    runbook_url?: string
+  }
+  endsAt?: string
+  fingerprint?: string
+  receivers?: [
+    {
+      name?: string
+    },
+  ]
+  startsAt?: string
+  status?: {
+    inhibitedBy?: Array<string>
+    silencedBy?: Array<string>
+    state?: string
+  }
+  updatedAt?: string
+  generatorURL?: string
+  labels?: {
+    alertname?: string
+    application?: string
+    businessUnit?: string
+    clusterName?: string
+    environment?: string
+    namespace?: string
+    productId?: string
+    prometheus?: string
+    queue_name?: string
+    severity?: string
+  }
+}
